[feature/#29] 로직 및 기능수정
