+++ src/main/java/com/oncha/oncha_web/feature/user/controller/MemberController.java
@@ -1,11 +1,19 @@
 package com.oncha.oncha_web.feature.user.controller;
 
+import com.amazonaws.services.kms.model.NotFoundException;
+import com.oncha.oncha_web.domain.user.model.Member;
+import com.oncha.oncha_web.domain.user.repository.MemberRepository;
+import com.oncha.oncha_web.feature.productBoard.model.ProductBoardDTO;
+import com.oncha.oncha_web.feature.user.model.MemberDTO;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 
 @Controller
+@RequiredArgsConstructor
 public class MemberController {
 
-
-
+    private final MemberRepository memberRepository;
+    
 }
